Decouple commissioner control cluster part 1
