+++ src/app/clusters/commissioner-control-server/app_config_dependent_sources.gni
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "commissioner-control-server.cpp",
-  "commissioner-control-server.h",
+  "CommissionerControlCluster.cpp",
+  "CodegenIntegration.h",
+  "Delegate.h",
+  "commissioner-control-server.h"
 ]
